Rename filenames to spell out abbreviations
