+++ src/graph/mod.rs
@@ -14,7 +14,7 @@ mod ford_fulkerson;
 mod graph_enumeration;
 mod heavy_light_decomposition;
 mod kosaraju;
-mod lee;
+mod lee_breadth_first_search;
 mod lowest_common_ancestor;
 mod minimum_spanning_tree;
 mod prim;
@@ -40,7 +40,7 @@ pub use self::ford_fulkerson::ford_fulkerson;
 pub use self::graph_enumeration::enumerate_graph;
 pub use self::heavy_light_decomposition::HeavyLightDecomposition;
 pub use self::kosaraju::kosaraju;
-pub use self::lee::lee;
+pub use self::lee_breadth_first_search::lee;
 pub use self::lowest_common_ancestor::{LowestCommonAncestorOffline, LowestCommonAncestorOnline};
 pub use self::minimum_spanning_tree::kruskal;
 pub use self::prim::{prim, prim_with_start};
